+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### [13.4.0](https://kaos.sh/ek/13.4.0)
 
+- `[system/sysctl]` Added new package for reading kernel parameters
 - `[strutil]` Added method `SqueezeRepeats`
 
 ### [13.3.5](https://kaos.sh/ek/13.3.5)